+++ packages/atproto_core/lib/src/sessions/session.g.dart
@@ -12,6 +12,7 @@ _$SessionImpl _$$SessionImplFromJson(Map<String, dynamic> json) =>
       handle: json['handle'] as String,
       email: json['email'] as String?,
       isEmailConfirmed: json['emailConfirmed'] as bool? ?? false,
+      emailAuthFactor: json['emailAuthFactor'] as bool? ?? false,
       accessJwt: json['accessJwt'] as String,
       refreshJwt: json['refreshJwt'] as String,
       didDoc: json['didDoc'] as Map<String, dynamic>?,
@@ -31,6 +32,7 @@ Map<String, dynamic> _$$SessionImplToJson(_$SessionImpl instance) {
 
   writeNotNull('email', instance.email);
   val['emailConfirmed'] = instance.isEmailConfirmed;
+  val['emailAuthFactor'] = instance.emailAuthFactor;
   val['accessJwt'] = instance.accessJwt;
   val['refreshJwt'] = instance.refreshJwt;
   writeNotNull('didDoc', instance.didDoc);