+++ packages/atproto/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release Note
 
+## v0.11.10
+
+- Add `authFactorToken` parameter on `createSession`. ([#1412](https://github.com/myConsciousness/atproto.dart/issues/1412))
+- Add `emailAuthFactor` property on `Session`. ([#1412](https://github.com/myConsciousness/atproto.dart/issues/1412))
+- Add `emailAuthFactor` property on `CurrentSession`. ([#1412](https://github.com/myConsciousness/atproto.dart/issues/1412))
+- Add `emailAuthFactor` parameter on `updateEmail`. ([#1412](https://github.com/myConsciousness/atproto.dart/issues/1412))
+
 ## v0.11.9
 
 - Added global `.headers` on `ATProto`. ([#1360](https://github.com/myConsciousness/atproto.dart/issues/1360))